CV_app/requirements copy.txt
accelerate==1.6.0
antlr4-python3-runtime==4.9.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
huggingface-hub==0.30.1
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opencv-contrib-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pip==23.0.1
portalocker==3.1.1
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
SAM-2==1.0
setuptools==65.5.0
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.51.0
typing_extensions==4.13.1
urllib3==2.3.0
